Separate out asserters
